feat: add banner principal
