+++ Front-end/src/App.jsx
@@ -1,15 +1,13 @@
-// import { useState } from 'react'
-// import reactLogo from './assets/react.svg'
-// import viteLogo from '/vite.svg'
-// import './App.css'
-
-import CardProduto from './Componentes/Card/card';
+import Banner from './components/Banner';
+import Card from './components/Card/card';
 
 function App() {
 
   return (
     <>
-      <CardProduto/>
+      <Banner />
+      <Card />
+      
     </>
   )
 }
