--- Front-end/src/Componentes/Card/card.css
@@ -1,63 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Manrope:wght@200..800&display=swap');
-
-.card {
-    padding: 10px;
-    width: 100%; 
-    height: auto;
-    text-align: center;
-    display: flex;
-    flex-direction: column;
-    row-gap: 6px;
-}
-
-.info-produto {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.nome-produto {
-    font-family: 'Manrope', sans-serif;
-    font-size: 16px;
-    color: #222222;
-    font-weight: 800;
-}
-
-.valor-produto {
-    font-family: 'Manrope', sans-serif;
-    font-size: 12px;
-    color: #4BAF47;
-    font-weight: 600;
-}
-
-
-@media (min-width: 600px) {
-    .card {
-        width: 90%;
-        padding: 15px;
-    }
-
-    .nome-produto {
-        font-size: 18px;
-    }
-
-    .valor-produto {
-        font-size: 14px;
-    }
-}
-
-@media (min-width: 1024px) {
-    .card {
-        width: 370px;
-        height: 458px;
-        padding: 20px;
-    }
-
-    .nome-produto {
-        font-size: 22px;
-    }
-
-    .valor-produto {
-        font-size: 16px;
-    }
-}
